refactor clean-code
