+++ src/components/Property.js
@@ -1,54 +1,21 @@
 import React from "react";
-import ObjectWrapProvider, { useObjectWrapProvider } from "./ObjectWrapContext";
+import ObjectWrapProvider from "./ObjectWrapContext";
 import "./Property.css";
 import ObjectWrap from "./ObjectWrap";
 import Switch from "./Switch";
 import { ReactComponent as TrashSvg } from "assets/trash.svg";
-import { schemaOperationsEnum } from "hooks/useSchema";
-import { defaultSchemaProperty } from "utils";
+import { usePropertyControls } from "hooks/usePropertyControls";
+import { validPropertyTypes } from "utils";
 
 const Property = ({ data }) => {
-  const { dispatch } = useObjectWrapProvider();
-
-  const updateProperty = (value) =>
-    dispatch({
-      type: schemaOperationsEnum.UPDATE,
-      payload: { key: data.id, value },
-    });
-
-  const renameTitleHandler = (event) =>
-    updateProperty({ title: event.target.value });
-
-  const changeInterfaceType = (event) => {
-    const changeTo = event.target.value;
-    if (changeTo === "object")
-      return updateProperty({ type: changeTo, children: {} });
-    updateProperty({ type: changeTo, children: null });
-  };
-
-  const addChild = () => {
-    let uuid = self.crypto.randomUUID();
-    console.log(uuid);
-    updateProperty({
-      type: "object",
-      children: {
-        ...data.children,
-        [uuid]: {
-          ...defaultSchemaProperty,
-          id: uuid,
-        },
-      },
-    });
-  };
-
-  const toggleIsRequired = (isChecked) =>
-    updateProperty({ isRequired: isChecked });
-
-  const removeProperty = () => {
-    dispatch({ type: schemaOperationsEnum.REMOVE, payload: data.id });
-  };
-
-  const observer = (updatedChild) => updateProperty({ children: updatedChild });
+  const {
+    renameTitleHandler,
+    changeInterfaceType,
+    addChild,
+    toggleIsRequired,
+    removeProperty,
+    observer,
+  } = usePropertyControls(data);
 
   return (
     <li>
@@ -65,10 +32,11 @@ const Property = ({ data }) => {
             className="type"
             onChange={changeInterfaceType}
           >
-            <option>string</option>
-            <option>object</option>
-            <option>number</option>
-            <option>boolean</option>
+            {Object.values(validPropertyTypes).map((type) => (
+              <option value={type} key={type}>
+                {type}
+              </option>
+            ))}
           </select>
         </section>
         <section className="controls">
